requirement.txt
fastapi
uvicorn
mangum
python-dotenv